Sync app quota to correct number, reset and disable rows quota
