+++ packages/server/src/utilities/usageQuota.js
@@ -5,24 +5,20 @@ const {
   generateNewUsageQuotaDoc,
 } = require("@budibase/auth/db")
 
-function getNewQuotaReset() {
-  return Date.now() + 2592000000
-}
-
 exports.Properties = {
-  ROW: "rows",
-  UPLOAD: "storage",
-  VIEW: "views",
-  USER: "users",
-  AUTOMATION: "automationRuns",
-  APPS: "apps",
-  EMAILS: "emails",
+  ROW: "rows", // mostly works - disabled - app / table deletion not yet accounted for
+  UPLOAD: "storage", // doesn't work yet
+  VIEW: "views", // doesn't work yet
+  USER: "users", // doesn't work yet
+  AUTOMATION: "automationRuns", // doesn't work yet
+  APPS: "apps", // works
+  EMAILS: "emails", // doesn't work yet
 }
 
-async function getUsageQuotaDoc(db) {
+exports.getUsageQuotaDoc = async db => {
   let quota
   try {
-    quota = await db.get(StaticDatabases.PLATFORM_INFO.docs.usageQuota)
+    quota = await db.get(StaticDatabases.GLOBAL.docs.usageQuota)
   } catch (err) {
     // doc doesn't exist. Create it
     quota = await db.post(generateNewUsageQuotaDoc())
@@ -45,15 +41,7 @@ exports.update = async (property, usage) => {
 
   try {
     const db = getGlobalDB()
-    const quota = await getUsageQuotaDoc(db)
-
-    // Check if the quota needs reset
-    if (Date.now() >= quota.quotaReset) {
-      quota.quotaReset = getNewQuotaReset()
-      for (let prop of Object.keys(quota.usageQuota)) {
-        quota.usageQuota[prop] = 0
-      }
-    }
+    const quota = await exports.getUsageQuotaDoc(db)
 
     // increment the quota
     quota.usageQuota[property] += usage
